Rename test_helpers.hpp to testing.hpp
The old name was confusingly similar to helpers.hpp. All 6 consumers
updated.
