+++ test/source/lob_coriolis_test.cpp
@@ -13,7 +13,7 @@
 #include "eng_units.hpp"
 #include "helpers.hpp"
 #include "lob/lob.hpp"
-#include "test_helpers.hpp"
+#include "testing.hpp"
 
 namespace tests {
