+++ test/source/lob_cwaj_test.cpp
@@ -14,7 +14,7 @@
 
 #include "eng_units.hpp"
 #include "lob/lob.hpp"
-#include "test_helpers.hpp"
+#include "testing.hpp"
 
 namespace tests {
