+++ test/source/lob_env_test.cpp
@@ -15,7 +15,7 @@
 #include "eng_units.hpp"
 #include "helpers.hpp"
 #include "lob/lob.hpp"
-#include "test_helpers.hpp"
+#include "testing.hpp"
 
 namespace tests {
